requirements.txt
PyOpenGL
pygame
PyOpenGL-accelerate
numpy
